modified model and api route to make the data fetch work
